Updated TicToc to use long to support longer timings.

+++ bi/share/src/bi/cache/AncestryCache.hpp
@@ -209,7 +209,7 @@ private:
   /**
    * Time taken for last write, in microseconds.
    */
-  int usecs;
+  long usecs;
 
   /**
    * Serialize.
